Move large test data variable into its own folder and use one flat top namespace instead of nested namespace.
BUG=195779890
