+++ tensorflow/lite/micro/kernels/testdata/conv_test_data.h
@@ -1,4 +1,4 @@
-/* Copyright 2020 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2021 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -19,14 +19,10 @@ limitations under the License.
 #include "tensorflow/lite/c/common.h"
 
 namespace tflite {
-namespace testing {
-namespace conv {
-extern const int8_t kInput1x32x32x3[];
-extern const int8_t kFilter8x3x3x3[];
-extern const int32_t kBiasQuantized8[];
-extern const int8_t kGoldenOutput1x16x16x8[];
-}  // namespace conv
-}  // namespace testing
+extern const int8_t kConvInput1x32x32x3[];
+extern const int8_t kConvFilter8x3x3x3[];
+extern const int32_t kConvBiasQuantized8[];
+extern const int8_t kConvGoldenOutput1x16x16x8[];
 }  // namespace tflite
 
 #endif  // TENSORFLOW_LITE_MICRO_KERNELS_CONV_TEST_DATA_H_
